+++ Assets/InfiniteScrollView/Scripts/Samples~/03_Grid/DemoVerticalGridCell.cs
@@ -6,7 +6,7 @@ public class DemoVerticalGridCell : InfiniteCell
 {
     public Text text;
 
-    public override void OnUpdate()
+    public override void OnRefresh()
     {
         RectTransform.sizeDelta = CellData.cellSize;
         text.text = CellData.index.ToString();